split up register into 2 forms
